ICP-13040 Change zigbee devices to report decimal temperature (#36278)
This accompanies hubcore and appengine-zigbee changes to allow the parsing and reporting of temperature values as floats.

+++ devicetypes/smartthings/ezex-temp-humidity-sensor.src/ezex-temp-humidity-sensor.groovy
@@ -76,7 +76,7 @@ def parse(String description) {
         }
     } else if (map.name == "temperature") {
         if (tempOffset) {
-            map.value = (int) map.value + (int) tempOffset
+            map.value = map.value + tempOffset
         }
         map.descriptionText = temperatureScale == 'C' ? '{{ device.displayName }} was {{ value }}°C' : '{{ device.displayName }} was {{ value }}°F'
         map.translatable = true
